+++ iOS/CIFilter.io/Filter List/Models/FilterCellModel.swift
@@ -10,25 +10,6 @@ import UIKit
 import ReactiveLists
 import SwiftUI
 
-extension TableCellViewModel where Self: View {
-    var registrationInfo: ViewRegistrationInfo {
-        return ViewRegistrationInfo(classType: HostingCell<Self>.self)
-    }
-
-    var cellIdentifier: String {
-        return String(describing: HostingCell<Self>.self)
-    }
-
-    var rowHeight: CGFloat? {
-        return UITableView.automaticDimension
-    }
-
-    func applyViewModelToCell(_ cell: UITableViewCell) {
-        guard let cell = cell as? HostingCell<Self> else { return }
-        cell.rootView = self
-    }
-}
-
 struct FilterCellModelView: TableCellViewModel, DiffableViewModel, View {
     var accessibilityFormat: CellAccessibilityFormat = "FilterCellModelView"
